Changes to the FIle
I edited this Zork file to include new exits and destinations for those exits. I also added Items such as a sign saying that it is pointing down and an ancient Egyptian Mummy that can be stored by the player. I also added numerous dead ends and descriptions for certain areas.